Conflicts:
	examples/src/Main.cpp
	framework/src/minko/render/IndexBuffer.cpp
	framework/src/minko/render/VertexBuffer.cpp

Former-commit-id: c8e98add4703dbd9f1a09cf727ec5d0ce85e8913
Former-commit-id: f738b524f74339c5f8e3fce273ceb1099a59c267
Former-commit-id: 9d9b91597e7532f025183ffc7d3180ca55c4d900

+++ framework/src/minko/render/IndexBuffer.cpp
@@ -25,8 +25,9 @@ using namespace minko::render;
 
 void
 IndexBuffer::upload()
-{
-	_id = _context->createIndexBuffer(_data.size());
+{   
+	if (_id == -1)
+    	_id = _context->createIndexBuffer(_data.size());
 	
 	_context->uploaderIndexBufferData(_id, 0, _data.size(), &_data[0]);
 }
@@ -34,6 +35,9 @@ IndexBuffer::upload()
 void
 IndexBuffer::dispose()
 {
-	_context->deleteIndexBuffer(_id);
-	_id = -1;
+    if (_id != -1)
+    {
+	    _context->deleteIndexBuffer(_id);
+	    _id = -1;
+    }
 }
